Build fix after 158223; make TileController use float for scales.
This matches what we do in other places, and fixes the constant
issue with exporting symbols that include CGFloat.

* WebCore.exp.in:
* platform/graphics/ca/mac/TileController.h:
(WebCore::TileController::scale):
* platform/graphics/ca/mac/TileController.mm:
(WebCore::TileController::setScale):


Canonical link: https://commits.webkit.org/141598@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@158230 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/ca/mac/TileController.h
@@ -59,8 +59,8 @@ public:
     void setNeedsDisplay();
     void setNeedsDisplayInRect(const IntRect&);
 
-    void setScale(CGFloat);
-    CGFloat scale() const { return m_scale; }
+    void setScale(float);
+    float scale() const { return m_scale; }
 
     bool acceleratesDrawing() const { return m_acceleratesDrawing; }
     void setAcceleratesDrawing(bool);
@@ -216,8 +216,8 @@ private:
     
     IntRect m_primaryTileCoverageRect; // In tile coords.
 
-    CGFloat m_scale;
-    CGFloat m_deviceScaleFactor;
+    float m_scale;
+    float m_deviceScaleFactor;
 
     TileCoverage m_tileCoverage;
     bool m_isInWindow;
